requirements.txt
# requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pandas
spacy
pyresparser
nltk
openpyxl
python-docx
flask
flask-cors
gunicorn    

